Remove http server from the test
