--- test/tests/node/every_http_loss.txt
@@ -1,1 +1,0 @@
-ok: all 120 events delivered, total POST attempts: 3
