+++ test/tests/node/every_loss.txt
@@ -1,0 +1,1 @@
+ok: all 120 events delivered
